[fixed] 전체적용내용수정

+++ avatar_crab2/app/src/main/java/piber/avatar_crab/presentation/data/HdaDataPoint.kt
@@ -3,8 +3,8 @@ package piber.avatar_crab.presentation.data
 data class HdaDataPoint(
     val hour: String,
     val min: Int,
-    val q1: Long,
-    val q2: Long,
-    val q3: Long,
+    val q1: Float,
+    val q2: Float,
+    val q3: Float,
     val max: Int
 )
